refactor(frontend): extract shared TopUpForm from WalletRefill
